create dedicated lib/app/stats.ex for metrics/stats related queries #397
